chore: rename components files to PascalCase
