+++ src/components/ui/header/CategoriesDropdown.jsx
