+++ src/components/ui/nav.jsx
@@ -1,10 +1,10 @@
 import { navLinks } from '@/lib/data'
 import { cn } from '@/lib/utils'
 import { NavLink } from 'react-router-dom'
-import UserAuth from '../user-auth'
-import { Button } from './button'
+import UserAuth from '../UserAuth'
+import { Button } from './Button'
 import { XIcon } from 'lucide-react'
-import { useNavContext } from '@/context/nav-context'
+import { useNavContext } from '@/context/NavContext'
 
 export function DesktopNav() {
   return (
